requirements.txt
streamlit==1.35.0
streamlit-option-menu==0.3.6
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
scikit-learn==1.5.0
Pillow==10.3.0
requests==2.32.3

